requirements.txt
flask
openai
werkzeug
pyyaml
python-dotenv
